+++ fireworks/core/firework.py
@@ -765,7 +765,7 @@ class Workflow(FWSerializable):
         # this should be done *before* additions
         if action.detours:
             for wf in action.detours:
-                new_updates = self._add_wf_to_fw(wf, fw_id, True)
+                new_updates = self.add_wf_to_fws(wf, [fw_id], pull_spec_mods=False, detour=True)
                 if len(set(updated_ids).intersection(new_updates)) > 0:
                     raise ValueError(
                         "Cannot use duplicated fw_ids when dynamically detouring workflows!")
@@ -774,7 +774,7 @@ class Workflow(FWSerializable):
         # add additional FireWorks
         if action.additions:
             for wf in action.additions:
-                new_updates = self._add_wf_to_fw(wf, fw_id, False)
+                new_updates = self.add_wf_to_fws(wf, [fw_id], pull_spec_mods=False, detour=False)
                 if len(set(updated_ids).intersection(new_updates)) > 0:
                     raise ValueError(
                         "Cannot use duplicated fw_ids when dynamically adding workflows!")
@@ -802,23 +802,24 @@ class Workflow(FWSerializable):
         # refresh the WF to get the states updated
         return self.refresh(fw_id, updated_ids)
 
-    def _add_wf_to_fw(self, wf, fw_id, detour):
+    def add_wf_to_fws(self, new_wf, fw_ids, pull_spec_mods=True, detour=False):
         """
         Internal method to add a workflow as a child to a Firework
+        Note: detours must have children that have STATE_RANK that is WAITING or below
 
-        :param wf: New Workflow to add
-        :param fw_id: id of the Firework on which to add the Workflow
-        :param detour: whether to add the children of the current Firework to
-         the Workflow's leaves
+        :param new_wf: (Workflow) New Workflow to add
+        :param fw_ids: ([int]) ids of the parent Fireworks on which to add the Workflow
+        :param pull_spec_mods: (bool) pull spec mods of COMPLETED parents
+        :param detour: (bool) add children of the current Firework to the Workflow's leaves
         :return: ([int]) list of Firework ids that were updated or new
         """
         updated_ids = []
 
-        root_ids = wf.root_fw_ids
-        leaf_ids = wf.leaf_fw_ids
+        root_ids = new_wf.root_fw_ids
+        leaf_ids = new_wf.leaf_fw_ids
 
-        for new_fw in wf.fws:
-            if new_fw.fw_id > 0:
+        for new_fw in new_wf.fws:
+            if new_fw.fw_id >= 0:  # note - this is also used later in the 'detour' code
                 raise ValueError(
                     'FireWorks to add must use a negative fw_id! Got fw_id: '
                     '{}'.format(
@@ -826,18 +827,37 @@ class Workflow(FWSerializable):
 
             self.id_fw[new_fw.fw_id] = new_fw  # add new_fw to id_fw
 
-            if new_fw.fw_id in leaf_ids:
-                if detour:
-                    self.links[new_fw.fw_id] = list(
-                        self.links[fw_id])  # add children of current FW to new FW
+            for fw_id in fw_ids:
+                if new_fw.fw_id in leaf_ids:
+                    if detour:
+                        # make sure all of these links are WAITING, else the DETOUR is not well defined
+                        ready_run = [(f >= 0 and Firework.STATE_RANKS[self.fw_states[f]] > 1) for f in self.links[fw_id]]
+                        if any(ready_run):
+                            raise ValueError("Detour option only works if all children of detours are not READY to run and have not already run")
+                        self.links[new_fw.fw_id] = [f for f in self.links[fw_id] if f >= 0]  # add children of current FW to new FW
+                    else:
+                        self.links[new_fw.fw_id] = []
                 else:
-                    self.links[new_fw.fw_id] = []
-            else:
-                self.links[new_fw.fw_id] = wf.links[new_fw.fw_id]
-            updated_ids.append(new_fw.fw_id)
+                    self.links[new_fw.fw_id] = new_wf.links[new_fw.fw_id]
+                updated_ids.append(new_fw.fw_id)
 
-        for root_id in root_ids:
-            self.links[fw_id].append(root_id)  # add the root id as my child
+        for fw_id in fw_ids:
+            for root_id in root_ids:
+                self.links[fw_id].append(root_id)  # add the root id as my child
+                if pull_spec_mods:  # re-apply some actions of the parent
+                    m_fw = self.id_fw[fw_id]  # get the parent FW
+                    m_launch = self._get_representative_launch(m_fw)  # get Launch of parent
+                    if m_launch:
+                        # pull spec update
+                        if m_launch.state == 'COMPLETED' and m_launch.action.update_spec:
+                            new_wf.id_fw[root_id].spec.update(m_launch.action.update_spec)
+                        # pull spec mods
+                        if m_launch.state == 'COMPLETED' and m_launch.action.mod_spec:
+                            for mod in m_launch.action.mod_spec:
+                                apply_mod(mod, new_wf.id_fw[root_id].spec)
+
+        for new_fw in new_wf.fws:
+            updated_ids = self.refresh(new_fw.fw_id, set(updated_ids))
 
         return updated_ids
 
@@ -847,7 +867,7 @@ class Workflow(FWSerializable):
 
         :param fw_id: (int) id of the Firework on which to perform the refresh
         :param updated_ids: ([int])
-        :return: ([int]) list of Firework ids that were updated
+        :return: (set(int)) list of Firework ids that were updated
         """
 
         updated_ids = updated_ids if updated_ids else set()  # these are the
@@ -877,21 +897,11 @@ class Workflow(FWSerializable):
         else:
             # my state depends on launch whose state has the highest 'score'
             # in STATE_RANKS
-            m_state = 'READY' if len(list(fw.launches)) == 0 else 'FIZZLED'
-            max_score = Firework.STATE_RANKS[m_state]
-            m_action = None
+            m_launch = self._get_representative_launch(fw)
+            m_state = m_launch.state if m_launch else 'READY'
+            m_action = m_launch.action if (m_launch and m_launch.state == "COMPLETED") else None
 
-            # TODO: pick the first launch in terms of end date that matches
-            # 'COMPLETED'; multiple might exist
-            for l in fw.launches:
-                if Firework.STATE_RANKS[l.state] > max_score:
-                    max_score = Firework.STATE_RANKS[l.state]
-                    m_state = l.state
-                    if m_state == 'COMPLETED':
-                        m_action = l.action
-
-            # This part is confusing and rare - report any FIZZLED parents if allow_fizzed
-            # allows us to handle FIZZLED jobs
+            # report any FIZZLED parents if allow_fizzed allows us to handle FIZZLED jobs
             if fw.spec.get('_allow_fizzled_parents'):
                 parent_fws = [self.id_fw[p].to_dict() for p in
                               self.links.parent_links.get(fw_id, []) if
@@ -1010,6 +1020,18 @@ class Workflow(FWSerializable):
 
     def _str_fw(self, fw_id):
         return self.id_fw[int(fw_id)].name + '--' + str(fw_id)
+
+    def _get_representative_launch(self, fw):
+        max_score = Firework.STATE_RANKS['ARCHIVED']  # state rank must be greater than this
+        m_launch = None
+
+        # TODO: if muliple COMPLETED launches, pick the first launch in terms of end date
+        for l in fw.launches:
+            if Firework.STATE_RANKS[l.state] > max_score:
+                max_score = Firework.STATE_RANKS[l.state]
+                m_launch = l
+
+        return m_launch
 
     @classmethod
     def from_dict(cls, m_dict):