+++ fireworks/tests/mongo_tests.py
@@ -9,7 +9,8 @@ import shutil
 import glob
 import unittest
 import time
-from fireworks.core.firework import Firework, Workflow
+from fireworks import explicit_serialize, FWAction
+from fireworks.core.firework import Firework, Workflow, FireTaskBase
 from fireworks.core.fworker import FWorker
 from fireworks.core.launchpad import LaunchPad, WFLock
 from fireworks.core.rocket_launcher import launch_rocket, rapidfire
@@ -52,6 +53,30 @@ def random_launch(lp_creds):
 def throw_error(msg):
     raise ValueError(msg)
 
+@explicit_serialize
+class MultipleDetourTask(FireTaskBase):
+
+    def run_task(self, fw_spec):
+        print('Running the Multiple Detour Task')
+        dt1 = Firework(ScriptTask.from_str('echo "this is intermediate job 1"'))
+        dt2 = Firework(ScriptTask.from_str('echo "this is intermediate job 2"'))
+        dt3 = Firework(ScriptTask.from_str('echo "this is intermediate job 3"'))
+        return FWAction(detours=[dt1, dt2, dt3])
+
+@explicit_serialize
+class UpdateSpecTask(FireTaskBase):
+    def run_task(self, fw_spec):
+        print('Running the Update Spec Task')
+        dt1 = Firework(ScriptTask.from_str('echo "this is dummy job 1"'))
+        return FWAction(update_spec={"dummy1": 1}, additions=[dt1])
+
+@explicit_serialize
+class ModSpecTask(FireTaskBase):
+    def run_task(self, fw_spec):
+        print('Running the Mod Spec Task')
+        return FWAction(mod_spec=[{"_push": {"dummy2": True}}])
+
+
 class MongoTests(unittest.TestCase):
 
     @classmethod
@@ -76,6 +101,14 @@ class MongoTests(unittest.TestCase):
 
     def setUp(self):
         self.old_wd = os.getcwd()
+
+    def tearDown(self):
+        self.lp.reset(password=None, require_password=False)
+        if os.path.exists(os.path.join('FW.json')):
+            os.remove('FW.json')
+        os.chdir(self.old_wd)
+        for i in glob.glob(os.path.join(MODULE_DIR, 'launcher*')):
+            shutil.rmtree(i)
 
     def test_basic_fw(self):
         test1 = ScriptTask.from_str("python -c 'print(\"test1\")'",
@@ -189,6 +222,18 @@ class MongoTests(unittest.TestCase):
         creds_array = [self.lp.to_dict()] * NCORES_PARALLEL_TEST
         p.map(random_launch, creds_array)
 
+    def test_multi_detour(self):
+        fw1 = Firework([MultipleDetourTask()], fw_id=1)
+        fw2 = Firework([ScriptTask.from_str('echo "DONE"')], parents=[fw1], fw_id=2)
+        self.lp.add_wf(Workflow([fw1, fw2]))
+        rapidfire(self.lp)
+        links = self.lp.get_wf_by_fw_id(1).links
+        self.assertEqual(set(links[1]), set([2, 3, 4, 5]))
+        self.assertEqual(set(links[2]), set([]))
+        self.assertEqual(set(links[3]), set([2]))
+        self.assertEqual(set(links[4]), set([2]))
+        self.assertEqual(set(links[5]), set([2]))
+
     def test_fworkerenv(self):
         t = DummyTask()
         fw = Firework(t)
@@ -252,6 +297,51 @@ class MongoTests(unittest.TestCase):
 
         self.assertEqual(self.lp.launches.count(), 1)
 
+    def test_add_wf_to_fws(self):
+        fw1 = Firework([UpdateSpecTask()])
+        fw2 = Firework([ModSpecTask()])
+        self.lp.add_wf(Workflow([fw1, fw2]))
+        self.assertEqual(self.lp.fireworks.count(), 2)
+        launch_rocket(self.lp, self.fworker)
+        launch_rocket(self.lp, self.fworker)
+        self.assertEqual(self.lp.launches.count(), 2)
+        self.assertEqual(self.lp.fireworks.count(), 3)  # due to detour
+
+        new_wf = Workflow([Firework([ModSpecTask()])])
+        self.lp.add_wf_to_fws(new_wf, [1, 2])
+        launch_rocket(self.lp, self.fworker)  # launch detour
+        launch_rocket(self.lp, self.fworker)  # launch new FW
+        launch_rocket(self.lp, self.fworker)  # dummy launch
+        new_fw = self.lp.get_fw_by_id(4)
+        self.assertEqual(new_fw.spec['dummy1'], 1)
+        self.assertEqual(new_fw.spec['dummy2'], [True])
+
+        self.assertEqual(self.lp.launches.count(), 4)
+        self.assertEqual(self.lp.fireworks.count(), 4)
+
+        new_wf = Workflow([Firework([ModSpecTask()])])
+        self.lp.add_wf_to_fws(new_wf, [4])
+        launch_rocket(self.lp, self.fworker)  # launch new FW
+        new_fw = self.lp.get_fw_by_id(5)
+        self.assertEqual(new_fw.spec['dummy2'], [True])
+
+        new_wf = Workflow([Firework([ModSpecTask()])])
+        self.assertRaises(ValueError, self.lp.add_wf_to_fws, new_wf, [4], detour=True)
+
+    def test_add_wf_to_fws_detour(self):
+        fw1 = Firework([ModSpecTask()], fw_id=1)
+        fw2 = Firework([ModSpecTask()], fw_id=2, parents=[fw1])
+        self.lp.add_wf(Workflow([fw1, fw2]))
+
+        new_wf = Workflow([Firework([ModSpecTask()])])
+        self.lp.add_wf_to_fws(new_wf, [1], detour=True)
+
+        launch_rocket(self.lp, self.fworker)
+        launch_rocket(self.lp, self.fworker)
+
+        self.assertEqual(self.lp.get_fw_by_id(2).spec['dummy2'], [True, True])
+
+
     def test_force_lock_removal(self):
         test1 = ScriptTask.from_str("python -c 'print(\"test1\")'", {'store_stdout': True})
         fw = Firework(test1, {"_dupefinder": DupeFinderExact()}, fw_id=1)
@@ -301,18 +391,6 @@ class MongoTests(unittest.TestCase):
         workflow_results=s.get_workflow_summary(time_field="updated_on")
         self.assertEqual((workflow_results[0]["_id"], workflow_results[0]["count"]), ("COMPLETED", 3))
 
-    def tearDown(self):
-        self.lp.reset(password=None, require_password=False)
-        if os.path.exists(os.path.join('FW.json')):
-            os.remove('FW.json')
-        os.chdir(self.old_wd)
-        for i in glob.glob(os.path.join(MODULE_DIR, 'launcher*')):
-            shutil.rmtree(i)
-
-    @classmethod
-    def tearDownClass(cls):
-        if cls.lp:
-            cls.lp.connection.drop_database(TESTDB_NAME)
 
 if __name__ == "__main__":
     unittest.main()